Conclusão do softDelete

+++ app/Services/Member/MemberService.php
@@ -3,14 +3,25 @@
 namespace App\Services\Member;
 
 use App\Repositories\Member\MemberRepositoryInterface;
+use App\Repositories\Transaction\TransactionRepositoryInterface;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Redirector;
 
 class MemberService implements MemberServiceInterface
 {
-    private $memberRepository;
-    public function __construct(MemberRepositoryInterface $memberRepository)
+    protected MemberRepositoryInterface $memberRepository;
+
+    protected TransactionRepositoryInterface $transactionRepository;
+
+    public function __construct(
+        MemberRepositoryInterface $memberRepository,
+        TransactionRepositoryInterface $transactionRepository
+    )
     {
         $this->memberRepository = $memberRepository;
+        $this->transactionRepository = $transactionRepository;
     }
 
     public function allMembers()
@@ -28,13 +39,29 @@ class MemberService implements MemberServiceInterface
         return $this->memberRepository->createMember($request);
     }
 
-    public function deleteMember(int $id)
+    public function deleteMember(int $id): Redirector|Application|RedirectResponse
     {
-        return $this->memberRepository->deleteMember($id);
+        $transaction = $this->transactionRepository->getTransactionByMemberId($id);
+
+        if ($transaction > 0) {
+
+            $this->memberRepository->softDeleteMember($id);
+
+            return redirect("/members")->with('success', "Cadastro removido com Sucesso!");
+        }
+
+        $this->memberRepository->deleteMember($id);
+
+        return redirect("/members")->with('success', "Cadastro removido com Sucesso!");
     }
 
     public function updateMember(Request $request, int $id)
     {
         return $this->memberRepository->updateMember($request, $id);
+    }
+
+    public function softDeleteMember(int $id)
+    {
+        return $this->memberRepository->softDeleteMember($id);
     }
 }
